refactor: 마이페이지 위치 변경
