--- src/pages/mypage/mypage.module.css
@@ -1,270 +1,0 @@
-.layout {
-  display: flex;
-  flex-wrap: nowrap;
-  align-items: flex-start;
-  justify-content: flex-start;
-  padding: 0rem;
-  min-height: 80vh;
-  background-color: var(--gray-FAFAFA);
-}
-.content {
-  padding: 2rem;
-
-  flex: 1;
-  max-width: 80rem;
-}
-
-.backWrapper {
-  margin-bottom: 2rem;
-}
-
-.backButton {
-  background: none;
-  border: none;
-  font-family: var(--font-family);
-  display: flex;
-  align-items: center;
-  cursor: pointer;
-  gap: 0.6rem;
-  outline: none;
-  box-shadow: none;
-  color: var(--black-333236);
-}
-
-.backIcon {
-  width: 1.6rem;
-  height: 1.6rem;
-}
-.backText {
-  font-family: var(--font-family);
-  font-size: 1.6rem;
-  line-height: 2.6rem;
-  font-weight: 500;
-}
-
-.cardGroup {
-  display: flex;
-  flex-direction: column;
-  gap: 2.4rem;
-}
-
-.card {
-  width: 66.9rem;
-  background: var(--white-FFFFFF);
-  padding: 3.2rem;
-  border-radius: 1.6rem;
-}
-
-.profileCard {
-  height: 36.6rem;
-}
-
-.passwordCard {
-  height: 46.6rem;
-}
-
-.sectionTitle {
-  font-family: var(--font-family);
-  font-size: 2.4rem;
-  line-height: 3.2rem;
-  font-weight: 700;
-  margin-bottom: 1.6rem;
-}
-
-.profileWrapper {
-  display: flex;
-  gap: 3.2rem;
-}
-
-.avatar {
-  width: 18.2rem;
-  height: 18.2rem;
-  border: 0.1rem var(--gray-D9D9D9);
-  border-radius: 1.6rem;
-  font-size: 3.2rem;
-  font-family: var(--font-family);
-  color: var(--violet-5534DhA);
-  display: flex;
-  background-color: #f5f5f5;
-  align-items: center;
-  justify-content: center;
-  cursor: pointer;
-  transition: background-color 0.2s ease;
-}
-.avatar img {
-  width: 100%;
-  height: 100%;
-  object-fit: cover;
-  border-radius: 1.6rem;
-}
-.avatar:hover {
-  background-color: var(--gray-FAFAFA);
-  border-color: var(--gray-9FA6B2);
-  color: var(--violet-5534DhA);
-}
-
-.profileForm,
-.passwordForm {
-  flex: 1;
-  display: flex;
-  flex-direction: column;
-}
-
-.profileForm label,
-.passwordForm label {
-  display: block;
-  margin-bottom: 0.4rem;
-  font-family: var(--font-family);
-  font-size: 1.6rem;
-  line-height: 2.6rem;
-  font-weight: 400;
-  color: var(--black-4B4B4B);
-}
-
-.profileForm input,
-.passwordForm input {
-  margin-bottom: 1.6rem;
-  font-size: 1.4rem;
-  font-family: var(--font-family);
-  border: 0.1rem solid var(--gray-D9D9D9);
-  border-radius: 0.8rem;
-}
-
-.profileForm input:focus,
-.passwordForm input:focus {
-  border: 0.1rem solid var(--violet-5534DhA);
-  outline: none;
-}
-
-.profileForm input {
-  padding: 1.5rem;
-}
-
-.profileForm input:last-of-type {
-  margin-bottom: 2.4rem;
-}
-.inputError {
-  border: 0.1rem solid var(--red-D6173A);
-}
-.passwordForm input {
-  padding: 1.5rem;
-}
-
-.errorMessage {
-  color: var(--red-D6173A);
-  font-size: 1.4rem;
-  margin-top: -1.6rem;
-}
-
-.saveButton,
-.changeButton {
-  font-family: var(--font-family);
-  font-weight: 600;
-  padding: 1.5rem;
-  border: none;
-  border-radius: 0.8rem;
-  color: var(--white-FFFFFF);
-  background-color: var(--violet-5534DhA);
-  cursor: pointer;
-  width: 100%;
-}
-
-.activeButton {
-  background-color: var(--violet-5534DhA);
-  cursor: pointer;
-}
-
-.inactiveButton {
-  background-color: var(--gray-D9D9D9);
-  cursor: not-allowed;
-}
-.saveButton {
-  padding: 1.5rem;
-}
-
-.changeButton {
-  padding: 1.5rem;
-  margin-top: 0.8rem;
-}
-.emailInput {
-  color: var(--gray-9FA6B2);
-  background-color: transparent;
-  border: none;
-}
-/* 태블릿 (max-width: 1023px) */
-@media (max-width: 1023px) {
-  .layout {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    min-height: 100vh;
-    padding: 0rem 0em;
-  }
-
-  .card {
-    width: 54.8rem;
-  }
-
-  .content {
-    padding: 7.5rem 18rem;
-  }
-
-  .profileCard {
-    height: 36.6rem;
-  }
-
-  .passwordCard {
-    height: 46.6rem;
-  }
-}
-
-/* 모바일 (max-width: 767px) */
-@media (max-width: 767px) {
-  .profileWrapper {
-    flex-direction: column;
-    align-items: flex-start;
-    gap: 5rem;
-  }
-  .content {
-    padding: 7.5rem 8rem;
-  }
-  .avatar {
-    width: 10rem;
-    height: 10rem;
-    font-size: 2.8rem;
-  }
-
-  .card {
-    width: 28.4rem;
-    padding: 2rem;
-  }
-  .sectionTitle {
-    font-family: var(--font-family);
-    font-size: 1.8rem;
-    line-height: 2.6rem;
-    font-weight: 700;
-    margin-bottom: 1.6rem;
-  }
-  .profileForm input,
-  .passwordForm input {
-    width: 25.2rem;
-    height: 5rem;
-  }
-
-  .saveButton,
-  .changeButton {
-    width: 25.2rem;
-    height: 5.4rem;
-  }
-
-  .profileCard {
-    height: 49.6rem;
-    padding: 2rem;
-  }
-  .backText {
-    font-size: 1.4rem;
-  }
-  .passwordCard {
-    height: 45.4rem;
-  }
-}
